Extract ILogEntry
