+++ app/src/lib/logging/log-entry.ts
@@ -1,0 +1,4 @@
+export interface ILogEntry {
+  level: 'error' | 'warn' | 'info' | 'debug'
+  readonly message: string
+}
